// src/app/backend/model/DiscoveredDevice.ts
/**
 * Magnusson Home Security
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: mange_magnusson@hotmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface DiscoveredDevice {
    serialNumber?: string;

    address?: string;

    modelName?: string;

    modelNumber?: string;

    lastDiscovered?: Date;

}